Cyclic elements, NamedNodeMap (#50)
Introduces a layer to elements to store common functionality like NamedNodeMap. In this we store a reference back up to the Node which owns the Element (Elements are always inside of Nodes because of our composition system which replaces the C++ inheritance model). Note that in order to do this we needed a new ConcreteNode constructor which can handle the cyclic nature. Fortunately, with the builders, this isn't so much an issue.

Also fleshes out NamedNodeMap to lay foundations for support for attributes.